REFACTOR done
