+++ master/src/main/scala/dpsort/master/MasterTaskServer.scala
@@ -77,7 +77,7 @@ private class MasterTaskServiceImpl extends MasterTaskServiceGrpc.MasterTaskServ
 
   override def reportTaskResult(request: TaskReportMsg): Future[ResponseMsg] = {
     logger.debug(s"Task report arrived")
-    TaskRunner.taskResultHandler( request )
+    TaskManager.taskResultHandler( request )
     val response = ResponseMsg( ResponseMsg.ResponseType.NORMAL )
     Future.successful( response )
   }
