+++ master/src/main/scala/dpsort/master/execution/Stage.scala
@@ -24,30 +24,28 @@
 
 package dpsort.master.execution
 
-import scala.collection.mutable.ListBuffer
-import scala.math
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
-import scala.util.Random
 
 import org.apache.logging.log4j.scala.Logging
 
 import dpsort.core.execution._
 import dpsort.core.network.TaskReportMsg
 import dpsort.core.network.TaskReportMsg.TaskResultType
-import dpsort.core.storage.PartitionMeta
-import dpsort.core.utils.IdUtils._
-import dpsort.core.utils.PartitionUtils._
-import dpsort.core.utils.SerializationUtils._
-import dpsort.master.{MasterContext, PartitionMetaStore, TaskRunner}
+import dpsort.master.{MasterContext, PartitionMetaStore, TaskManager}
 import dpsort.master.MasterConf._
-import dpsort.master.TaskRunner._
+import dpsort.master.TaskManager._
 
 
 object StageExitStatus extends Enumeration {
   val SUCCESS, FAILURE = Value
 }
 
+/*
+ * With the indent to make workers stupid,
+ * the stage is hidden to the worker.
+ * Worker only executes TaskContexts.
+ */
 trait Stage extends Logging {
 
   protected val stageTaskSet: TaskSet = genTaskSet()
@@ -77,7 +75,7 @@ trait Stage extends Logging {
 
   def executeAndWaitForTermination(): StageExitStatus.Value = {
     logger.info(s"Register stage ${this.toString}.")
-    val registerResult = TaskRunner.registerStage(this )
+    val registerResult = TaskManager.registerStage(this )
     if( !registerResult ) {
       logger.error(s"Cannot register stage to TaskRunner")
       return StageExitStatus.FAILURE
@@ -98,202 +96,5 @@ trait Stage extends Logging {
       }
     }
     stageExitStatus
-  }
-}
-
-class EmptyStage extends Stage {
-  override def toString: String = "EmptyStage"
-  override protected def genTaskSet(): TaskSet = {
-    val taskSeq: Iterable[BaseTask] = {
-      for (i <- 0 to 9;
-           wid <- PartitionMetaStore.getWorkerIds)
-        yield new EmptyTask(genNewTaskID, wid, TaskStatus.WAITING, Unit, Unit)
-    }
-    logger.info(s"${taskSeq.size} task(s) generated")
-    new TaskSet( Random.shuffle( taskSeq ) ) // for fair scheduling
-  }
-}
-
-class TerminateStage extends Stage {
-  override def toString: String = "TerminateStage"
-
-  override protected def genTaskSet(): TaskSet = {
-    val taskSeq: Iterable[BaseTask] = {
-      // only 1 termination for each worker is possible.
-      var outPartIdx = 1;
-      def genOutPartName(idx: Int) = "partition."+idx.toString
-      for ( wid <- PartitionMetaStore.getWorkerIds )
-        yield {
-          val resultPart: PartitionMeta = PartitionMetaStore.getPartitionList(wid).head
-          val outFileName = genOutPartName(outPartIdx)
-          outPartIdx += 1
-          new TerminateTask(genNewTaskID, wid, TaskStatus.WAITING, resultPart.pName, outFileName )
-        }
-    }
-    new TaskSet( Random.shuffle( taskSeq ) ) // for fair scheduling
-  }
-
-}
-
-class GenBlockStage extends Stage {
-  override def toString: String = "GenBlockStage"
-
-  override protected def genTaskSet(): TaskSet = {
-    val taskSeq: Iterable[BaseTask] = {
-      for ( wid <- PartitionMetaStore.getWorkerIds )
-        yield {
-          val parts = PartitionMetaStore.getPartitionList( wid )
-          parts.map( pMeta => {
-            val offsets: Array[(Int, Int)] = blockOffsetsGenerator( pMeta, get("dpsort.master.blockSizeInLines").toInt )
-            val outPNames: Array[String] = {1 to offsets.size}.toArray.map( _ => genNewPartID )
-            assert( offsets.size == outPNames.size )
-            new GenBlockTask( genNewTaskID, wid, TaskStatus.WAITING, pMeta.pName, outPNames, offsets )
-          }).toArray
-        }
-    }.flatten
-    new TaskSet( Random.shuffle( taskSeq ) )
-  }
-
-  override def taskResultHandler(taskRes: TaskReportMsg): Unit = {
-    if( taskRes.taskResult == TaskResultType.SUCCESS ) {
-      val task = stageTaskSet.getTask( taskRes.taskId )
-      val wid = task.getWorkerID
-      PartitionMetaStore.delPartitionMeta( wid, task.inputPartition.head )
-      task.outputPartition.foreach(
-        outPart => {
-          PartitionMetaStore.genAndAddPartitionMeta( wid, outPart )
-        }
-      )
-    }
-    super.taskResultHandler( taskRes )
-  }
-
-}
-
-class LocalSortStage extends Stage {
-
-  override def toString: String = "LocalSortStage"
-
-  override protected def genTaskSet(): TaskSet = {
-    val taskSeq: Iterable[BaseTask] = {
-    for (wid <- PartitionMetaStore.getWorkerIds)
-      yield {
-        val parts = PartitionMetaStore.getPartitionList(wid)
-        parts.map(pMeta => {
-          new LocalSortTask(genNewTaskID, wid, TaskStatus.WAITING, pMeta.pName, genNewPartID )
-        }).toArray
-      }
-    }.flatten
-    new TaskSet( Random.shuffle( taskSeq ) )
-}
-
-  override def taskResultHandler(taskRes: TaskReportMsg): Unit = {
-    if( taskRes.taskResult == TaskResultType.SUCCESS ) {
-      val task = stageTaskSet.getTask( taskRes.taskId )
-      val wid = task.getWorkerID
-      PartitionMetaStore.delPartitionMeta( wid, task.inputPartition.head )
-      task.outputPartition.foreach(
-        outPart => PartitionMetaStore.genAndAddPartitionMeta( wid, outPart )
-      )
-    }
-    super.taskResultHandler( taskRes )
-  }
-
-}
-
-class SampleKeyStage extends Stage {
-  override def toString: String = "SampleKeyStage"
-
-  override protected def genTaskSet(): TaskSet = {
-    val recordsCnt = MasterContext.recordsCount
-    val sampleRatio = math.min( get("dpsort.master.maxSampleSize").toFloat / recordsCnt.toFloat
-                                ,get("dpsort.master.maxSampleRatio").toFloat )
-    logger.info(s"sample ratio is set to ${sampleRatio}")
-    val taskSeq: Iterable[BaseTask] = {
-      for (wid <- PartitionMetaStore.getWorkerIds)
-        yield {
-          val parts = PartitionMetaStore.getPartitionList(wid)
-          parts.map(pMeta => {
-            new SampleKeyTask(genNewTaskID, wid, TaskStatus.WAITING, pMeta.pName, genNewPartID, sampleRatio )
-          }).toArray
-        }
-    }.flatten
-    new TaskSet( Random.shuffle( taskSeq ) )
-  }
-
-  override def taskResultHandler(taskRes: TaskReportMsg): Unit = {
-    val sampledKeys = deserializeByteStringToObject( taskRes.serializedTaskResultData )
-      .asInstanceOf[Array[Array[Byte]]]
-    MasterContext.registryLock.lock()
-    try {
-      MasterContext.sampledKeys ++= sampledKeys
-    } catch {
-      case e: Throwable => {
-        logger.error("failed to aggregate sampled values from the task output. bypassing the samples")
-      }
-    } finally {
-        MasterContext.registryLock.unlock()
-    }
-    super.taskResultHandler( taskRes )
-  }
-}
-
-class PartitionAndShuffleStage extends Stage {
-  override def toString: String = "PartitionAndShuffleStage"
-
-  override protected def genTaskSet(): TaskSet = {
-    val pFunc = MasterContext.partitionFunction
-    val taskSeq: Iterable[BaseTask] = {
-      for ( wid <- PartitionMetaStore.getWorkerIds )
-        yield {
-          val parts = PartitionMetaStore.getPartitionList( wid )
-          parts.map( pMeta => {
-            val outPNames: Array[String] = {1 to PartitionMetaStore.getWorkerIds.size}.toArray.map( _ => genNewPartID() )
-            new PartitionAndShuffleTask( genNewTaskID, wid, TaskStatus.WAITING, pMeta.pName, outPNames, pFunc )
-          }).toArray
-        }
-    }.flatten
-    new TaskSet( Random.shuffle( taskSeq ) )
-  }
-
-  override def taskResultHandler(taskRes: TaskReportMsg): Unit = {
-    if( taskRes.taskResult == TaskResultType.SUCCESS ) {
-      val task = stageTaskSet.getTask( taskRes.taskId )
-      val wid = task.getWorkerID
-      PartitionMetaStore.delPartitionMeta( wid, task.inputPartition.head )
-      task.outputPartition.zipWithIndex.foreach(
-        outpartIdx => PartitionMetaStore.genAndAddPartitionMeta( outpartIdx._2 + 1, outpartIdx._1 )
-      )
-    }
-    super.taskResultHandler( taskRes )
-  }
-
-}
-
-class MergeStage extends Stage {
-  override def toString: String = "MergeStage"
-
-  override protected def genTaskSet(): TaskSet = {
-    val taskSeq: Iterable[BaseTask] = {
-      for ( wid <- PartitionMetaStore.getWorkerIds )
-        yield {
-          val parts: ListBuffer[PartitionMeta] = PartitionMetaStore.getPartitionList( wid )
-          parts.grouped(2).toList.filter( pair => pair.size == 2).map(
-            pair => new MergeTask( genNewTaskID, wid, TaskStatus.WAITING, pair.map( el => el.pName ).toArray, genNewPartID )
-          ).toArray
-        }
-    }.flatten
-    new TaskSet( Random.shuffle( taskSeq ) )
-  }
-
-  override def taskResultHandler(taskRes: TaskReportMsg): Unit = {
-    if( taskRes.taskResult == TaskResultType.SUCCESS ) {
-      val task = stageTaskSet.getTask( taskRes.taskId )
-      val wid = task.getWorkerID
-      PartitionMetaStore.delPartitionMeta( wid, task.inputPartition(0) )
-      PartitionMetaStore.delPartitionMeta( wid, task.inputPartition(1) )
-      PartitionMetaStore.genAndAddPartitionMeta( wid, task.outputPartition(0) )
-    }
-    super.taskResultHandler( taskRes )
   }
 }
